pyproject.toml
[tool.poetry]
name = "urlshortener"
version = "0.1.0"
description = ""
authors = ["jakubk <jakramp96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.1"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
extend-exclude = ['migrations/']

[tool.ruff]
line-length = 110
extend-exclude = ['migrations/']